[FIX] web_tree_many2one_clickable: Charge assets after list renderer component
Before these changes, if a module with many2one fields was loaded before this one, an error occurred saying that the TreeMany2oneClickableButton component could not be found.

To reproduce the error: Install dms and web_tree_many2one_clickable, go to Documents > Files, switch to tree view, and display the Directory field.

By forcing the assets to load right after the web module, we prevent the other module from loading first, thus avoiding this error.