[IMP] Added medium devices Bootstrap condition on control panel CSS
Small devices (smartphones and some tablets) possess a different layout for the control panel, in which the searchbar is hidden, then revealed upon pressing a button which then takes the screen.
Since our module causes interferences with the small devices layout, as it was thought with medium devices and up (tablets and desktop/laptops) in mind, we add a medium devices condition (max-width of >= 768 px) to apply the CSS to wider screens only.

We use this implementation since this is also how the desktop and mobile control panel CSS is differenciated in the web module (see web/static/src/search/control_panel/control_panel.scss file).