[MIG] Reimplemented group percentages JS module
To reimplement the module with the new framework, we patched the ListRenderer class inheriting Component, adding a sum of all of the values of numeric columns variable, and a method to get any given group's percentage compared to the column total sum.
To insert the percentage in the template, we inherit the groupRow template, and add the percentage within the aggregate value cell.

When some values are big enough, they cannot be displayed entierly within a same line. As such, we add a linebreak within the cells and display the percentage underneath the value to show it clearly.