[IMP] Restore table layout for aggregated lists
In the freezeColumnWidth method (which is used whenever the window is resized), the table element used to display the list views is forced into a fixed layout, which can lead to text overflow with the added percentages next to the values.
As such, whenever the list is grouped, we set the table layout to auto, to reduce the probability of overflow.