[FIX] web_group_by_percentage: Added exclusion of non-field columns
In some list/tree views, there are non-field columns, for instance buttons to handle an action on a group of lines.
Consequently, with this module installed, it's impossible to group lines in certain views, if the grouped view has a button,
as using isNumericColumn, which tries to access the column in the view's fields, throws an Exception.

As such, we exclude non-field columns from the computation of grouped columns percentages.